part3/requirements.txt
flask-jwt-extended==4.4.4
pyjwt==2.6.0
flask-bcrypt
flask-restx

